[ADD] Funciona el search de alias de los usuarios por ajax
